Renamed sheen to stp. Fixed test.
