+++ rj_gameplay/.gitignore
@@ -1,1 +1,1 @@
-/sheen.egg-info/
+/stp.egg-info/
