+++ rj_gameplay/README.md
@@ -1,2 +1,2 @@
-# Sheen - Gameplay for RJ RoboCup SSL
+# stp - Gameplay for RJ RoboCup SSL
 This is a library containing the rewrite of the RJ RoboCup gameplay library.
