--- rj_gameplay/sheen/utils/__init__.py
@@ -1,1 +1,0 @@
-import sheen.utils.edict
