+++ rj_gameplay/stp/skill/seeker.py
@@ -1,6 +1,6 @@
 from abc import ABC, abstractmethod
 
-import sheen.skill as skill
+import stp.skill as skill
 
 
 class ISeeker(skill.ISkill, ABC):
