+++ rj_gameplay/tests/play/test_play.py
@@ -1,1 +1,1 @@
-import sheen.play as play
+import stp.play as play
