+++ rj_gameplay/tests/role/test_cost.py
@@ -1,7 +1,7 @@
 import numpy as np
 
-import sheen.role.cost as cost
-from sheen.game_state import Robot
+import stp.role.cost as cost
+from stp.game_state import Robot
 
 
 def test_constant() -> None:
@@ -25,8 +25,9 @@ def test_constant() -> None:
 def test_distance_to_pt() -> None:
     """Tests cost.distance_to_pt for both the same robot and different robot cases."""
     switch_cost = 0.5
-    target_pt = np.ndarray([1.0, 2.0])
-    cost_fn = cost.distance_to_pt(target_pt, switch_cost)
+    saturate_dist = 1.0
+    target_pt = np.array([1.0, 2.0])
+    cost_fn = cost.distance_to_pt(target_pt, saturate_dist, switch_cost)
 
     robot1 = Robot(1, np.array([1.0, 2.0, 0]), np.array([0, 0, 0]), False)
     robot2 = Robot(2, np.array([2.0, 3.0, 0]), np.array([0, 0, 0]), False)
@@ -38,7 +39,7 @@ def test_distance_to_pt() -> None:
     assert cost_same == 0.0
 
     cost_robot2: float = cost_fn(robot2, None)
-    assert cost_robot2 == np.sqrt(2)
+    assert cost_robot2 == 1.0
 
     cost_robot2_switch: float = cost_fn(robot2, robot1)
     assert cost_robot2_switch == 1.0
