+++ rj_gameplay/tests/tactic/test_pass_or_shoot.py
@@ -1,4 +1,4 @@
-import sheen.tactic.pass_or_shoot as pass_or_shoot
+import stp.tactic.pass_or_shoot as pass_or_shoot
 
 
 def test_pass_or_shoot():
